refactor remove.conf translation code to new module
